daemon: move txo `jsonrpc_` methods to an individual module
